server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pyserial>=3.5
pywebpush>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
